backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pydantic-settings
python-dotenv
httpx
soco
plexapi
pytest
pytest-asyncio
pyyaml
